+++ core-tests/test.hs
@@ -1,9 +1,12 @@
+{-# LANGUAGE DeriveDataTypeable #-}
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Tasty.Runners
 import Test.Tasty.Options
 import Test.Tasty.Patterns.Types
 import Data.Maybe
+import Data.Typeable
+import Options.Applicative
 import Resources
 import Timeouts
 import Dependencies
@@ -22,6 +25,7 @@ mainGroup = do
     , testDependencies
     , patternTests
     , awkTests_
+    , optionMessagesTests
     ]
 
 -- | 'patternTests' are not supposed to test every awk feature; that's the
@@ -70,3 +74,29 @@ tt =
   where
     -- trivial HUnit test
     t s = testCase s (return ())
+
+-- | Test the behavior of the warning messages that are generated from
+-- command-line argument parsers. See #270.
+optionMessagesTests :: TestTree
+optionMessagesTests = testGroup "OptionMessages"
+  [ testCase "JankyOption generates warning messages" $
+      length warnings @?= 2
+  ]
+  where
+    (warnings, _) = optionParser [Option (Proxy :: Proxy JankyOption)]
+
+data JankyOption = MkJankyOption Int Int
+  deriving Typeable
+instance IsOption JankyOption where
+  defaultValue   = MkJankyOption 27 42
+  parseValue     = \_ -> return defaultValue
+  optionName     = return "janky-option"
+  optionHelp     = return "This is an incredibly janky option. For testing purposes only!"
+  -- This implementation of optionCLParser does two shameful things that
+  -- beget warning messages:
+  --
+  -- 1. It uses Options.Applicative.value.
+  -- 2. It defines two options instead of just one.
+  optionCLParser = MkJankyOption
+    <$> option auto (value 27)
+    <*> option auto (value 42)