+++ core/CHANGELOG.md
@@ -4,6 +4,16 @@ Changes
 Version ?
 ---------
 
+* In addition to a `Parser`, `optionParser` and `suiteOptionParser` now return
+  a `[String]` representing warning messages:
+  * A warning is emitted if an `IsOption` instance defines multiple options in
+    the implementation of `optionCLParser`.
+  * An warning is emitted if an `IsOption` instance's `optionCLParser`
+    implementation assigns a default value (e.g., with
+    `Options.Applicative.value`), as this interferes with `tasty`'s ability to
+    read environment variable arguments.
+* `IsOption` has a new method `showDefaultValue` for customizing how
+  `defaultValue`s are rendered in the `--help` output.
 * Drop support for GHC versions older than 8.0
 
 Version 1.2.3